refactor: remove redux
