+++ client/src/store/actionTypes.js
@@ -1,7 +1,9 @@
-export const SET_INITIAL_STATES = "set-initial-states";
+export const SET_MY_DATA = "set-my-data";
 export const UPDATE_MY_PROFILE = "update-my-profile";
 
 export const REMOVE_DRAFT = "remove-draft";
+
+export const INCREMENT_VIEW = "increment-view";
 
 // OLD
 export const ITEM_ADDED = "item-added";
