+++ server/src/model/Review.js
@@ -3,12 +3,7 @@ import { reqStr, reqNum, defaultDate } from "./values.js";
 
 const { Schema, model } = mongoose;
 
-const iReview = new Schema({
-  id: reqStr,
-  reviewId: reqStr,
-});
-
-const reviewMe = new Schema({
+const details = new Schema({
   reviewerId: reqStr,
   date: defaultDate,
   rating: reqNum,
@@ -18,10 +13,10 @@ const reviewMe = new Schema({
 
 const review = new Schema({
   id: reqStr,
-  reviewMe: [reviewMe],
+  privateId: reqStr,
+  reviews: [details],
   totalRating: { type: Number, default: 0 },
   numOfReviews: { type: Number, default: 0 },
-  iReview: [iReview],
 });
 
 const Review = model("reviews", review);
