+++ server/src/model/index.js
@@ -1,7 +1,7 @@
 import Restriction from "./Restriction.js";
-import Listing from "./Listing.js";
 import Account from "./Account.js";
 import Favourite from "./Favourite.js";
 import Review from "./Review.js";
+import IReview from "./IReview.js";
 
-export { Restriction, Listing, Account, Favourite, Review };
+export { Restriction, Account, Favourite, Review, IReview };
